requirements.txt
flask
tensorflow
Pillow
boto3
waitress
